# docs/swagger.yaml
definitions:
...
  users.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  users.user:
    properties:
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /users/:
    get:
      consumes:
      - application/json
      description: Returns a list of users
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/users.user'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/users.ErrorResponse'
      summary: Returns users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a new user with the provided data
      parameters:
      - description: New user data
        in: body
        name: newUser
        required: true
        schema:
          $ref: '#/definitions/users.user'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.user'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/users.ErrorResponse'
      summary: Creates a new user
      tags:
      - users
  /users/user/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a user with the specified ID
      parameters:
      - description: User ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.user'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/users.ErrorResponse'
      summary: Deletes a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Returns a single user by ID
      parameters:
      - description: User ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.user'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/users.ErrorResponse'
      summary: Returns single user
      tags:
      - users
swagger: "2.0"